requirements.txt
flask
flask-mysql
flask-sqlalchemy
sqlalchemy-migrate